backend/api_proxy/requirements.txt
fastapi
httpx
python-consul
starlette
sse_starlette
PyJWT